chimera/requirements.txt
# Core dependencies
pydantic>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0

# Database
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0

# Web3 and blockchain
web3>=6.0.0
eth-account>=0.10.0
eth-utils>=2.0.0
eth-abi>=4.0.0

# Async support
aiohttp>=3.9.0
websockets>=12.0

# AWS integration
boto3>=1.28.0

# Logging and monitoring
structlog>=23.0.0
prometheus-client>=0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
